requirements.txt
openai
chromadb
pypdf
numpy
python-dotenv
rank-bm25